fix(secrets): serialize environment deletion with scoped writes
